Move NMSUtils
