+++ spigot/src/main/java/net/apartium/cocoabeans/spigot/inventory/InventoryNMSUtils.java
@@ -8,7 +8,7 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.apartium.cocoabeans.spigot.utils;
+package net.apartium.cocoabeans.spigot.inventory;
 
 import org.bukkit.Bukkit;
 
@@ -18,7 +18,8 @@ import java.lang.reflect.Method;
 /**
  * @deprecated Will be replaced with an actually normal util later on. Not for use outside the library itself
  */
-public class NMSUtils {
+@Deprecated
+/* package-private */ class InventoryNMSUtils {
 
     private static String version;
     private static final Class<?> craftItemStack;
